catalog-update: replace window width counting to up level
